refactor: 循環依存3件を解消（定義移動+importパス変更のみ）
修正1: server-manager.ts → global-init.ts の値import解消
- getEnvironmentStatus等を utils/environment-status.ts に移動
- server-manager.ts は同層の environment-status.ts から直接import
- global-init.ts のre-exportは environment-status.ts に切り替え

修正2: constants/messages.ts → setup/setup-ui.ts の値import解消
- SetupStep interface を constants/setup-steps.ts に移動
- messages.ts は同層の setup-steps.ts から直接import
- setup-ui.ts は constants/setup-steps.ts からre-export

ロジック変更なし。ビルド・テスト通過確認済み。
